Presumed Production Ready Update

+++ components/buttons/ticket/delete-ticket.tsx
@@ -8,45 +8,53 @@ import axios from "axios";
 import { toast } from "../../ui/sonner";
 
 
-const DeleteTicket: React.FC<HtmlHTMLAttributes<HTMLButtonElement> & {ticketId: string; variant?: any; onSuccess?: Callback; onFailure?: Callback }> = ({
-    children, className, ticketId, variant, onSuccess, onFailure, ...props
+const DeleteTicket: React.FC<HtmlHTMLAttributes<HTMLButtonElement> & {
+    ticketId: string;
+    variant?: any;
+    onPending?: Callback;
+    onSuccess?: Callback;
+    onFailure?: Callback
+}> = ({
+    children, className, ticketId, variant, onPending, onSuccess, onFailure, ...props
 }) => {
-    const actor = useAuthenticatedUser();
+        const actor = useAuthenticatedUser();
 
-    const handleDelete: React.MouseEventHandler<HTMLButtonElement> = async (ev: MouseEvent) => {
-        if ( !ev.isDefaultPrevented() ) {
-            ev.preventDefault();
-            ev.stopPropagation();
-        }
-        const url = Api.server + Api.endpoints.admin.singleTicket.replace(':id', ticketId);
-        try {
-            const response = await axios.delete(url, { 
-                headers: { Authorization: `Bearer ${actor?.token}`}
-            });
-            
-            if ( response.status == 200 ) {
-                if (onSuccess) {
-                    onSuccess(response.data);
+        const handleDelete: React.MouseEventHandler<HTMLButtonElement> = async (ev: MouseEvent) => {
+            if (!ev.isDefaultPrevented()) {
+                ev.preventDefault();
+                ev.stopPropagation();
+            }
+            if (!confirm('Are you sure you want to delete this ticket?')) {
+                return;
+            }
+            toast(<div className="text-destructive">Deleting ticket</div>)
+            onPending && onPending(null);
+
+            const url = Api.server + Api.endpoints.admin.singleTicket.replace(':id', ticketId);
+            try {
+                const response = await axios.delete(url, {
+                    headers: { Authorization: `Bearer ${actor?.token}` }
+                });
+
+                if (response.status == 200) {
+                    onSuccess && onSuccess(response.data);
+                    toast(<div className="text-green-800">Ticket deleted successfully</div>);
                 }
-                toast('Ticket deleted successfully');
+            } catch (error) {
+                onFailure && onFailure(error);
+                toast(<div className="text-destructive">Sorry, but your request could not be completed. Please try again later.</div>);
             }
-        } catch (error) {
-            if ( onFailure ) {
-                onFailure(error);
-            }
-            toast('Sorry, but your request could not be completed. Please try again later.');
         }
-    }
 
-    return (
-        <Button 
-            className={ cn(className) } 
-            type="button"
-            variant={ variant || 'default' } 
-            {...props} onClick={ handleDelete }>
-            { children || 'Delete' } <BiTrash size={ 20 } />
-        </Button>
-    )
-}
+        return (
+            <Button
+                className={cn(className)}
+                type="button"
+                variant={variant || 'default'}
+                {...props} onClick={handleDelete}>
+                {children || 'Delete'} <BiTrash size={20} />
+            </Button>
+        )
+    }
 
 export default DeleteTicket;
